refactor: standardize file naming to use kebab-case for consistency
