+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "sports-brain-buzz",
-    "version": "17.0.0",
+    "version": "17.0.1",
     "private": true,
     "type": "module",
     "description": "A web application for sports enthusiasts using React and TypeScript.",
