+++ src/_services/useNextQuestion.ts
@@ -1,5 +1,5 @@
 import { getNextQuestion } from '../_application/getNextQuestion.ts';
-import { useQuizStorage } from './store/storageAdapter.ts';
+import { useQuizStorage } from './store/storage-adapter.ts';
 
 export const useNextQuestion = () => {
     const quizStorage = useQuizStorage();
